Finished the skeleton of the rest service. SOAP login is added.

+++ backend/src/main/java/tcss556/services/models/requests/create/CreateUserData.java
@@ -17,9 +17,9 @@ public class CreateUserData {
     @NotNull
     private String password;
     private UserGroup group;
-    private int privilege;
-    private int floor;
-    private double location_x;
-    private double location_y;
+    private Long privilege;
+    private Integer floor;
+    private Double location_x;
+    private Double location_y;
 
 }
